+++ src/main/groovy/com/netflix/gradle/plugins/packaging/validation/SystemPackagingAttributeValidator.groovy
@@ -1,0 +1,6 @@
+package com.netflix.gradle.plugins.packaging.validation
+
+interface SystemPackagingAttributeValidator {
+    boolean validate(String attribute)
+    String getErrorMessage(String attribute)
+}